client/requirements.txt
gradio==5.45.0
python-dotenv==1.1.1
